# backend/src/main/java/io/swagger/api/RegisterApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.19).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.annotations.*;
import io.swagger.model.RegistrationData;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-04-05T23:15:12.065+02:00")

@Validated
@Api(value = "register", description = "the register API")
@RequestMapping(value = "")
public interface RegisterApi {

	@ApiOperation(value = "Create an account", nickname = "createAccount", notes = "", tags = {"auth",})
	@ApiResponses(value = {
			@ApiResponse(code = 201, message = "User registered successfully."),
			@ApiResponse(code = 400, message = "User's passwords do not match. The user's email is used in the application. The password is not strong enough.")})
	@RequestMapping(value = "/register",
			produces = {"application/json"},
			consumes = {"application/json"},
			method = RequestMethod.POST)
	ResponseEntity<Void> createAccount(@ApiParam(value = "User data containing email and passwords", required = true) @Valid @RequestBody RegistrationData body);

}