# ism_swagger_RecyclEasy_218598.yaml
swagger: "2.0"
info:
  description: "The 'RecyclEasy' is a web application, that helps people find places near them, where they can dispose recyclabe garbage. As well as, publish advertisements related to recyclable garbage managment and disposal."
  version: "1.0.0"
  title: "RecycleEasy"
schemes:
  - "https"
  - "http"
paths:
  /register:
    post:
      tags:
        - "auth"
      summary: "Create an account"
      description: ""
      operationId: "createAccount"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User data containing email and passwords"
          required: true
          schema:
            $ref: "#/definitions/RegistrationData"
      responses:
        201:
          description: "User registered successfully."
        400:
          description: "User's passwords do not match. The user's email is used in the application. The password is not strong enough."
...
  /resetPassword:
    put:
      tags:
        - "auth"
      summary: "Reset password to the system"
      description: ""
      operationId: "resetPassowrd"
      consumes:
        - "application/x-www-form-urlencoded"
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "formData"
          type: string
          name: "email"
          description: "Account email"
          required: true
      responses:
        200:
          description: "Reset link has been sent to user's email."
        400:
          description: "User email not found."
...
securityDefinitions:
  recycleasy_auth:
    type: "oauth2"
    authorizationUrl: "http://find.your.team.com/login"
    flow: "implicit"
    scopes:
      write: "modify data"
      read: "read data"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  RegistrationData:
    type: object
    required:
      - email
      - password
      - repeatPassword
      - acceptRules
    properties:
      email:
        type: string
      password:
        type: string
      repeatPassword:
        type: string
      acceptRules:
        type: boolean
...